+++ src/ts/main.ts
@@ -97,6 +97,7 @@ export { AdControlBar, AdControlBarConfig } from './components/ads/AdControlBar'
 export { AdMessageLabel } from './components/ads/AdMessageLabel';
 export { AdClickOverlay } from './components/ads/AdClickOverlay';
 export { AdCounterLabel } from './components/ads/AdCounterLabel';
+export { PauseAdStatusOverlay } from './components/ads/PauseAdStatusOverlay';
 export { PlaybackSpeedSelectBox } from './components/settings/PlaybackSpeedSelectBox';
 export { HugeReplayButton } from './components/buttons/HugeReplayButton';
 export { BufferingOverlay, BufferingOverlayConfig } from './components/overlays/BufferingOverlay';